commit without trash
