--- .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Mon Jan 02 19:34:52 EET 2023
-gradle.version=7.4
