--- app/src/main/java/com/example/messengerAndroid/ui/contactsViewModel/SimpleCallBack.kt
@@ -1,25 +1,0 @@
-package com.example.messengerAndroid.ui.contactsViewModel
-
-import androidx.recyclerview.widget.ItemTouchHelper
-import androidx.recyclerview.widget.RecyclerView
-
-abstract class SwipeToDeleteCallback : ItemTouchHelper.Callback() {
-    override fun getMovementFlags(
-        recyclerView: RecyclerView,
-        viewHolder: RecyclerView.ViewHolder
-    ): Int {
-        val swipeFlags = ItemTouchHelper.LEFT
-        return makeMovementFlags(0, swipeFlags)
-    }
-
-    // This feature hasn`t implemented in this project yet
-    override fun onMove(
-        recyclerView: RecyclerView,
-        viewHolder: RecyclerView.ViewHolder,
-        target: RecyclerView.ViewHolder
-    ): Boolean {
-        return false
-    }
-
-
-}
