+++ app/src/main/java/com/example/messengerAndroid/ui/myContacts/contactsViewModel/Getters.kt
@@ -1,0 +1,13 @@
+package com.example.messengerAndroid.ui.myContacts.contactsViewModel
+
+import com.example.messengerAndroid.data.contactsRepository.contactModel.User
+
+fun ContactsViewModel.getUserPosition(user: User): Int {
+    return contactsLiveData.value?.indexOf(user) ?: -1
+}
+
+fun ContactsViewModel.getUser(position: Int): User {
+    return contactsLiveData.value?.toMutableList()?.get(position)!!
+}
+
+
